Extract `MarkdownToReact` component into own file
