+++ website/src/components/util.js
@@ -1,10 +1,7 @@
-import React, { Fragment, useEffect, useState } from 'react'
+import React, { Fragment } from 'react'
 import { Parser as HtmlToReactParser } from 'html-to-react'
 import siteMetadata from '../../meta/site.json'
 import { domain } from '../../meta/dynamicMeta.mjs'
-import remarkPlugins from '../../plugins/index.mjs'
-import { serialize } from 'next-mdx-remote/serialize'
-import { MDXRemote } from 'next-mdx-remote'
 
 const htmlToReactParser = new HtmlToReactParser()
 
@@ -77,34 +74,6 @@ export function isEmptyObj(obj) {
  */
 export function htmlToReact(html) {
     return htmlToReactParser.parse(html)
-}
-
-/**
- * Convert raw Markdown to React
- * @param {String} markdown - The Markdown markup to convert.
- * @param {Object} [remarkReactComponents] - Optional React components to use
- *  for HTML elements.
- * @returns {Node} - The converted React elements.
- */
-export function MarkdownToReact({ markdown }) {
-    const [mdx, setMdx] = useState(null)
-
-    useEffect(() => {
-        const getMdx = async () => {
-            setMdx(
-                await serialize(markdown, {
-                    parseFrontmatter: false,
-                    mdxOptions: {
-                        remarkPlugins,
-                    },
-                })
-            )
-        }
-
-        getMdx()
-    }, [markdown])
-
-    return mdx ? <MDXRemote {...mdx} /> : <></>
 }
 
 /**
